Moved num embedded out of object's flags into it's own short int field.

+++ StandardLib/ArrayCommonObject.cpp
@@ -612,7 +612,7 @@ int CArrayCommonObject::RemapPointerTos(CEmbeddedObject* pcOld, CEmbeddedObject*
 //
 //
 //////////////////////////////////////////////////////////////////////////
-CEmbeddedObject* CArrayCommonObject::GetEmbeddedObject(int iIndex)
+CEmbeddedObject* CArrayCommonObject::GetEmbeddedObject(unsigned short int iIndex)
 {
 	if (iIndex == 0)
 	{
